BUG: Remove unused setWorkflow / workflow methods
Removing the methods solve the memory leak as a side effect. The leaks
occurs because (1) the module widget references both the steps and the workflow
and (1) the steps themselves also reference the workflow.
Since the workflow is referenced in both object, the steps are not
destructed and we can observe the leak.

To solve the issue while keeping the method around, the saferef python
module should be used to create a weak reference to the workflow.

See http://www.na-mic.org/Bug/view.php?id=2588
Fixes slicer isse 2588