BUG: reset the transform for the cropped ROI
The bug could be due to the fact that in the past, BRAINSFit registration would
not automatically set the transform for the moving volume, but this apparently
has been fixed. Cropped followup ROI volume is cloned from the followup volume,
which has the transform set up, which causes the misalignment on visualization.

Resolves issues #15 and #14.